fix: hrr가 정수로만 제약되는점 해결

+++ sjf.c
@@ -161,7 +161,7 @@ int hrn(process *pro,int n,int *solution){
 			if(pro[i].arrive_t <= time && pro[i].complete != 1){
 
 				// calculate the response ratio
-				temp = (pro[i].burst_t + (time - pro[i].arrive_t)) / pro[i].burst_t;
+				temp = (double)(pro[i].burst_t + (time - pro[i].arrive_t)) / pro[i].burst_t;
 				printf("현재 %d 번째 계산중인 temp%f\n",i,temp);
 				// 방금 계산한 새로운 프로세스가 이전에 가장 컸던 hrr 보다 크다면 업데이트 해준다.
 				if(hrr < temp){
